+++ src/components.d.ts
@@ -63,6 +63,7 @@ export namespace Components {
     'imageSupport': ExternalImageSupport;
     'version': string;
   }
+  interface PieEmbed {}
   interface PiePlayer {
     /**
     * Simulates a correct response for the item. This property will only have this effect if the `hosted` property is false and player is running client-side-only.
@@ -129,6 +130,12 @@ declare global {
     new (): HTMLPieAuthorElement;
   };
 
+  interface HTMLPieEmbedElement extends Components.PieEmbed, HTMLStencilElement {}
+  var HTMLPieEmbedElement: {
+    prototype: HTMLPieEmbedElement;
+    new (): HTMLPieEmbedElement;
+  };
+
   interface HTMLPiePlayerElement extends Components.PiePlayer, HTMLStencilElement {}
   var HTMLPiePlayerElement: {
     prototype: HTMLPiePlayerElement;
@@ -160,6 +167,7 @@ declare global {
   };
   interface HTMLElementTagNameMap {
     'pie-author': HTMLPieAuthorElement;
+    'pie-embed': HTMLPieEmbedElement;
     'pie-player': HTMLPiePlayerElement;
     'pie-preview-control': HTMLPiePreviewControlElement;
     'pie-preview-layout': HTMLPiePreviewLayoutElement;
@@ -212,6 +220,7 @@ declare namespace LocalJSX {
     'onModelUpdated'?: (event: CustomEvent<any>) => void;
     'version'?: string;
   }
+  interface PieEmbed {}
   interface PiePlayer {
     /**
     * Simulates a correct response for the item. This property will only have this effect if the `hosted` property is false and player is running client-side-only.
@@ -283,6 +292,7 @@ declare namespace LocalJSX {
 
   interface IntrinsicElements {
     'pie-author': PieAuthor;
+    'pie-embed': PieEmbed;
     'pie-player': PiePlayer;
     'pie-preview-control': PiePreviewControl;
     'pie-preview-layout': PiePreviewLayout;
@@ -298,6 +308,7 @@ declare module "@stencil/core" {
   export namespace JSX {
     interface IntrinsicElements {
       'pie-author': LocalJSX.PieAuthor & JSXBase.HTMLAttributes<HTMLPieAuthorElement>;
+      'pie-embed': LocalJSX.PieEmbed & JSXBase.HTMLAttributes<HTMLPieEmbedElement>;
       'pie-player': LocalJSX.PiePlayer & JSXBase.HTMLAttributes<HTMLPiePlayerElement>;
       'pie-preview-control': LocalJSX.PiePreviewControl & JSXBase.HTMLAttributes<HTMLPiePreviewControlElement>;
       'pie-preview-layout': LocalJSX.PiePreviewLayout & JSXBase.HTMLAttributes<HTMLPiePreviewLayoutElement>;